Move to static mem

+++ tracker/src/aiming/utm.c
@@ -1,38 +1,38 @@
 #include "utm.h"
 #include <math.h>
 
-utm_coord_t latlon_to_utm(double lat, double lon) {
+utm_coord_t latlon_to_utm(float lat, float lon) {
     utm_coord_t utm;
     
     // Calculate zone for central meridian
     int zone = (int)((lon + 180.0) / 6.0) + 1;
     
     // Convert to radians
-    double lat_rad = lat * DEG_TO_RAD;
-    double lon_rad = lon * DEG_TO_RAD;
+    float lat_rad = lat * DEG_TO_RAD;
+    float lon_rad = lon * DEG_TO_RAD;
     
     // Central meridian for the zone
-    double lon0 = ((zone - 1) * 6.0 - 180.0 + 3.0) * DEG_TO_RAD;
-    double lon_diff = lon_rad - lon0;
+    float lon0 = ((zone - 1) * 6.0 - 180.0 + 3.0) * DEG_TO_RAD;
+    float lon_diff = lon_rad - lon0;
     
     // Pre-calculate trig functions
-    double sin_lat = sin(lat_rad);
-    double cos_lat = cos(lat_rad);
-    double tan_lat = tan(lat_rad);
+    float sin_lat = sin(lat_rad);
+    float cos_lat = cos(lat_rad);
+    float tan_lat = tan(lat_rad);
     
     // Calculate N (radius of curvature in prime vertical)
-    double N = WGS84_A / sqrt(1 - WGS84_E2 * sin_lat * sin_lat);
+    float N = WGS84_A / sqrt(1 - WGS84_E2 * sin_lat * sin_lat);
     
     // Calculate T and C
-    double T = tan_lat * tan_lat;
-    double C = WGS84_E2 * cos_lat * cos_lat / (1 - WGS84_E2);
-    double A = cos_lat * lon_diff;
+    float T = tan_lat * tan_lat;
+    float C = WGS84_E2 * cos_lat * cos_lat / (1 - WGS84_E2);
+    float A = cos_lat * lon_diff;
     
     // Calculate M (meridional arc)
-    double e4 = WGS84_E2 * WGS84_E2;
-    double e6 = e4 * WGS84_E2;
+    float e4 = WGS84_E2 * WGS84_E2;
+    float e6 = e4 * WGS84_E2;
     
-    double M = WGS84_A * ((1 - WGS84_E2/4 - 3*e4/64 - 5*e6/256) * lat_rad
+    float M = WGS84_A * ((1 - WGS84_E2/4 - 3*e4/64 - 5*e6/256) * lat_rad
                 - (3*WGS84_E2/8 + 3*e4/32 + 45*e6/1024) * sin(2*lat_rad)
                 + (15*e4/256 + 45*e6/1024) * sin(4*lat_rad)
                 - (35*e6/3072) * sin(6*lat_rad));
